Add licensing comment block to LoginForm.cs
Introduce a new comment block at the beginning of LoginForm.cs
to provide legal and licensing information. This includes
copyright information for the year 2025 attributed to Anas Yusuf,
and states that OpenGYM is licensed under the GNU General Public
License (GPL) version 3 or later. The comment also clarifies
that OpenGYM is distributed without any warranty and provides
a link to the GPL license for more details.